requirement.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
cryptography==3.4.8
cycler==0.10.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-bootstrap5==2.1.2
django-paypal==1.1.1
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.6
wincertstore==0.2
